🟢 airport delete invalid airport id

+++ src/airport/airport.service.ts
@@ -45,6 +45,9 @@ export class AirportService {
 
     async delete(id: string) {
         const airport: AirportEntity = await this.airportRepository.findOne({ where: { id } });
+        if (!airport)
+            throw new BusinessLogicException("The airport with the given id was not found", BusinessError.NOT_FOUND);
+
         await this.airportRepository.remove(airport);
     }
 }
